v1.0

+++ declarations/constants/voidstrikerConstants.h
@@ -3,21 +3,21 @@
 //	Weapon facings
 //	F = fore, A = aft, S = starboard, P = port
 //	Seeking weapons have no facing
-const int8_t WF_F = 0;
-const int8_t WF_A = 1;
-const int8_t WF_S = 2;
-const int8_t WF_P = 3;
-const int8_t WF_FA = 4;
-const int8_t WF_FP = 5;
-const int8_t WF_FS = 6;
-const int8_t WF_AP = 7;
-const int8_t WF_AS = 8;
-const int8_t WF_PS = 9;
-const int8_t WF_FPA = 10;
-const int8_t WF_FSA = 11;
-const int8_t WF_FPS = 12;
-const int8_t WF_APS = 13;
-const int8_t WF_FPSA = 14;
-const int8_t WF_SEEKING = 15;
+const int WF_F = 0;
+const int WF_A = 1;
+const int WF_S = 2;
+const int WF_P = 3;
+const int WF_FA = 4;
+const int WF_FP = 5;
+const int WF_FS = 6;
+const int WF_AP = 7;
+const int WF_AS = 8;
+const int WF_PS = 9;
+const int WF_FPA = 10;
+const int WF_FSA = 11;
+const int WF_FPS = 12;
+const int WF_APS = 13;
+const int WF_FPSA = 14;
+const int WF_SEEKING = 15;
 
 //	Damage results
