+++ declarations/functions/habitabilityModifiersBreathable.h
@@ -1,2 +1,2 @@
 #include <cstdint>
-int8_t habitabilityModifiersBreathable(char pressureCategory, bool breathable);
+int habitabilityModifiersBreathable(char pressureCategory, bool breathable);
