+++ declarations/functions/habitabilityModifiersBreathableClimate.h
@@ -1,2 +1,2 @@
 #include <cstdint>
-int8_t habitabilityModifiersBreathableClimate(char climateType, bool breathable);
+int habitabilityModifiersBreathableClimate(char climateType, bool breathable);
