+++ declarations/functions/habitabilityModifiersNonBreathable.h
@@ -1,3 +1,3 @@
 #include <cstdint>
 #include "../structures/atmosphericComposition_t.h"
-int8_t habitabilityModifiersNonBreathable(char pressureCategory, atmosphericComposition_t atmosphereComposition);
+int habitabilityModifiersNonBreathable(char pressureCategory, atmosphericComposition_t atmosphereComposition);
