+++ declarations/structures/voidstrikerShip_t.h
@@ -8,34 +8,34 @@
 struct voidstrikerShip_t
 {
 	std::string shipName;
-	int8_t numAction;
-	int8_t thrustRatioNumerator;
-	int8_t thrustRatioDenominator;
-	int8_t thrusterType;
-	int8_t targetSize;
+	int numAction;
+	int thrustRatioNumerator;
+	int thrustRatioDenominator;
+	int thrusterType;
+	int targetSize;
 	int16_t constructionValue;
 	std::string className;
-	int8_t troops;
-	int8_t hangarCraftNum;
-	int8_t jumpSpeed;
-	int8_t jumpActions;
+	int troops;
+	int hangarCraftNum;
+	int jumpSpeed;
+	int jumpActions;
 	std::string raceName;
 	int16_t cargo;
-	int8_t burnNum;
-	int8_t shieldType;
-	int8_t shieldNum;
-	int8_t ewNum;
+	int burnNum;
+	int shieldType;
+	int shieldNum;
+	int ewNum;
 
 //	Armor table
-	int8_t armorTable[10][20];
+	int armorTable[10][20];
 //	Structure table
 	char structureTable[5][20];
 //	Frame table
-	int8_t frameTable[2][20];
+	int frameTable[2][20];
 
 //	Weapons
 //	First column is weapon type and second is quantity
-	int8_t weaponTable[16][2];
+	int weaponTable[16][2];
 //	Facing column for weapon table
-	int8_t facingTable[16];
+	int facingTable[16];
 }
